requirements.txt
apache-airflow
kafka-python
cassandra-driver
spark 
pyspark==3.5.1